Speed up svd load by dropping unused parent field
The `parent` element is included as a reference in the various SVD
types, and is used to compute a register's absolute memory offset based
on the parent peripheral.

Copying the entire `parent` object is expensive, because we copy the
entire peripheral tree that's been parsed at that point.

Since we only use the offset value, replace the reference to the entire
`parent` object with just the offset value.

To verify this patch worked correctly, I used the following snippet to
output the peripheral register addresses, and diffed the output before
and after this patch:

```python
svd = SVDFile(f)
for periph in svd.peripherals.values():
    print(periph.name)
    for reg in periph.registers.values():
        print("{} {}".format(reg.name, reg.address()))
```

I also removed the `SVDPeripheralRegisterField` refactor_parent function
and usage, since the parent information is never used by the fields
(only the `self.offset` value is used, and that's part of the Field SVD
element itself).