cmdebug: svd: handle enumeratedValue.value in binary format
According to the SVD specification:
https://www.keil.com/pack/doc/CMSIS/SVD/html/svd_xml_conventions_gr.html#Constants
, an enumeratedValue.value can be a binary value: it starts with the
character '#'. It is widely used in Freescale SVDs.

This patch corrects the current value handling. Instead of considering #string like
hexstring, it converts it into a binary value.

Besides let's use the 'index' variable for self.enum index instead of
converting it from the string 'value' again.