Only load pkg_resources optionally.
This import is only used to find installed SVD files as part of
cmsis-svd.

But when using this project as part of a minimal Python installation
(such as if you try to use the Zephyr SDK on openSUSE Tumbleweed) there
is no pip, nor setuptools, nor ability to install cmsis-svd.

So don't fail to load, just ignore the cmsis_svd loading if the import
fails.