LoadSVD: Load SVD from ZIP file instead of pkg_resources
This changes the logic with `svd_load <VENDOR> <MODEL>` to not search in
the pkg_resources of Python module `cmsis_svd.data` but instead search
in a ZIP file at `$CMSIS_SVD_ZIPFILE` if that environment variable is
defined, otherwise it looks `$XDG_CACHE_HOME/cmdebug/cmsis-svd-data.zip`
with `$XDG_CACHE_HOME` falling back to `~/.cache` if not found in the
environment.

As of 2025-11-09, the `HEAD` of https://github.com/cmsis-svd/cmsis-svd-data
is 5.8 GiB in size uncompressed, or 254 MiB when zipped. Personally, I
do not perceive any additional latency by having to unzip the chosen
SVD file first.